+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Change Log
 
-### backend-metal-0.5.1 (26-02-2020)
+### backend-vulkan-0.5.2 (01-04-2020)
+  - fix support for `AMD_NEGATIVE_VIEWPORT_HEIGHT`
+
+### backend-metal-0.5.1 (26-03-2020)
   - fix debug assertion for the index buffer range
   - fix `NDC_Y_FLIP` feature
 
-### backend-vulkan-0.5.1 (26-02-2020)
+### backend-vulkan-0.5.1 (26-03-2020)
   - fix debug color markers
   - fix detection of the `MirrorClamp` mode